% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.imbalanced.R
\name{is.imbalanced}
\alias{is.imbalanced}
\title{is.imbalanced}
\usage{
is.imbalanced(x)
}
\arguments{
\item{x}{vector.}
}
\value{
data.frame of error metrics.
}
\description{
examines whether a binary or multinomial feature is imbalanced
}
\author{
E. F. Haghish
}
