% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlim.load.R, R/mlim.load2.R
\name{mlim.load}
\alias{mlim.load}
\title{loads mlim object from mlim file}
\usage{
mlim.load(file)

mlim.load(file)
}
\arguments{
\item{file}{path to the mlim file on the disk}
}
\value{
object of class mlim

list
}
\description{
loads mlim object and creates a list

loads mlim object and creates a list
}
\author{
E. F. Haghish
}
