% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlim.postimpute.R
\name{mlim.postimpute}
\alias{mlim.postimpute}
\title{carries out postimputation to minimize imputation error}
\usage{
mlim.postimpute(
  data,
  preimputed.data,
  algos = c("ELNET", "GBM", "RF", "DL", "Ensemble"),
  ...
)
}
\arguments{
\item{data}{data.frame with missing values}

\item{preimputed.data}{data.frame. if you have used another software for missing
data imputation, you can still optimize the imputation
by handing the data.frame to this argument, which will
bypass the "preimpute" procedure.}

\item{algos}{character vector of algorithms to be used for imputation. the default
is 'c("ELNET", "GBM", "RF", "DL", "Ensemble")'.}

\item{...}{arguments to be passed to mlim}
}
\value{
imputed data.frame
}
\description{
this is a function for carrying out postimputation on datasets
             that are already imputed, either by mlim or other software.
}
\author{
E. F. Haghish
}
