% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlim.preimpute.R
\name{mlim.preimpute}
\alias{mlim.preimpute}
\title{Preimputation of missing values}
\usage{
mlim.preimpute(data, preimpute = "random", seed = NULL)
}
\arguments{
\item{data}{data.frame containing missing values}

\item{preimpute}{character. Specify the algorithm for preimputation.
Supported options are "mm" (median/mode replacement),
"random" for random sampling from available data.}

\item{seed}{integer. Random-number seed used by Random Forest or random
sampling. The default is \code{NULL}.}
}
\value{
imputed data.frame
}
\description{
Initializes missing values before the iterative mlim imputation procedure.
Missing values can be initialized using median/mode, random sampling,
from the observed values of each variable, or Random Forest imputation (experimental).
}
\details{
NOTE:
====

This file includes "RF" for randomforest preimputation. Such an approach can
inflate relationships between the features and thus, this method is not documented
and is not recommended. It is maintained in the code for research purpose only.
}
\examples{
\dontrun{
data(iris)

# add 10\% stratified missing values to one factor variable
irisNA <- iris
irisNA$Species <- mlim.na(irisNA$Species, p = 0.1, stratify = TRUE, seed = 2022)

# run the default Median/Model preimputation
MLIM <- mlim.preimpute(irisNA, preimpute = "mm")
mlim.error(MLIM, irisNA, iris) #check the preimputation error

# Random-sampling preimputation
RANDOM <- mlim.preimpute(irisNA, preimpute = "random", seed = 2022)
}
}
\author{
E. F. Haghish
}
