% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlim.prepare.R
\name{mlim.prepare}
\alias{mlim.prepare}
\title{prepares an object of mlim class for postimputation}
\usage{
mlim.prepare(data, imputed, md.log)
}
\arguments{
\item{mlim.object}{object of class mlim to continue the}

\item{save}{imputed data.frame with missing values. for example, if you use
"mlim.preimpute" function for preimputation, the returned data.frame
can be handed to this argument.}

\item{algorithm}{character. specify the algorithm for preimputation. the
supported options are "kNN", "ranger", "missForest", and "mm"
(mean-mode replacement). "ranger" and "missForest" both
implement Random Forest procedures and are generally
recommended. For very large datasets use "kNN".}
}
\value{
imputed data.frame
}
\description{
mlim is a computation intensive procedure. therefore it makes
             sense to secure the procedure by automating the process of
             saving the results of imputation. therefore, if R crashes, you
             need to follow the imputation from where it was stopped, instead
             of starting over. this process is carried out with "mlim.postimpute"
             function. however, "mlim.prepare" function only prepares the
             "mlim" class object needed for post imputation.
}
\author{
E. F. Haghish
}
