% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlim.shuffle.R
\name{mlim.shuffle}
\alias{mlim.shuffle}
\title{generate multiple imputation datasets from a single imputation}
\usage{
mlim.shuffle(df, imputed, m, report = NULL)
}
\arguments{
\item{df}{original dataframe with NAs}

\item{imputed}{single imputed dataset of the class 'mlim'. the dataset must have
been imuted with mlim R package to include the imputation metrics
needed for the computation}

\item{m}{integer, specifying number of multiple imputations. the default value is
1, carrying out a single imputation.}

\item{report}{filename. if a filename is specified (e.g. report = "mlim.md"), the \code{"md.log"} R
package is used to generate a Markdown progress report for the
imputation. the format of the report is adopted based on the
\code{'verbosity'} argument. the higher the verbosity, the more
technical the report becomes. if verbosity equals "debug", then
a log file is generated, which includes time stamp and shows
the function that has generated the message. otherwise, a
reduced markdown-like report is generated. default is NULL.}
}
\value{
list, including multiple imputed datasets
}
\description{
mlim estimates Root Mean Squared Error (RMSE), which can provide
             a hint about the standard deviation of the error. if normal
             distribution of the error is assumed, stochastic imputed values
             can be generated, where each predicted value is considered the
             'mean' of the normal distribution and the standard deviation is
             assumed to be equal to 'RMSE'.
}
\author{
E. F. Haghish
}
