% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlim.summarize.R
\name{mlim.summarize}
\alias{mlim.summarize}
\title{Summarize mlim imputation error}
\usage{
mlim.summarize(data)
}
\arguments{
\item{data}{An object returned by \code{mlim()}.}
}
\value{
For a single imputation, a data frame containing
variable names and RMSE values. For multiple imputation,
a data frame containing one RMSE column per imputation.
}
\description{
Summarizes the best accepted
cross-validation RMSE for each imputed variable.
}
\examples{
\dontrun{
data(iris)

irisNA <- iris
irisNA$Species <- mlim.na(
  irisNA$Species,
  p = 0.1,
  stratify = TRUE,
  seed = 2022
)

imp <- mlim(irisNA)
mlim.summarize(imp)
}
}
\author{
E. F. Haghish
}
