% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charity.R
\docType{data}
\name{charity}
\alias{charity}
\title{some items about attitude towards charity}
\format{
A data frame with 832 rows and 5 variables:
\describe{
  \item{ta1}{Charitable Organizations More Effective}
  \item{ta2}{Degree of Trust}
  \item{ta3}{Charitable Organizations Honest/Ethical}
  \item{ta4}{Role Improving Communities}
  \item{ta5}{Job Delivering Services}
}
}
\source{
\url{https://www.stata.com/}
}
\usage{
charity
}
\description{
A dataset containing likert-scale items about attitude towards charity
}
\keyword{datasets}
