% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{mlim.error}
\alias{mlim.error}
\title{imputation error}
\usage{
mlim.error(imputed, incomplete, complete)
}
\arguments{
\item{imputed}{the imputed dataframe}

\item{incomplete}{the dataframe with missing values}

\item{complete}{the original dataframe with no missing values}
}
\description{
calculates NRMSE, missclassification rate, and miss-ranking
             absolute mean distance, scaled from 0 to 1, where 1 means
             maximum distance between the actual rank of a level with the
             imputed level.
}
\author{
E. F. Haghish
}
