% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlim.mids.R
\name{mlim.mids}
\alias{mlim.mids}
\title{prepare "mids" class object}
\usage{
mlim.mids(mlim, X)
}
\arguments{
\item{mlim}{array of class "mlim", returned by "mlim" function}

\item{seed}{integer. a random seed number for reproducing the result (recommended)}
}
\description{
takes "mlim" object and prepares a "mids" class for data analysis with
             multiple imputation.
}
\examples{

\donttest{
# adding stratified NA to an atomic vector
x <- as.factor(c(rep("M", 100), rep("F", 900)))
table(mlim.na(x, p=.5, stratify = TRUE))

# adding unstratified NAs to all variables of a data.frame
data(iris)
mlim.na(iris, p=0.5, stratify = FALSE, seed = 1)

# or add stratified NAs only to factor variables, ignoring other variables
mlim.na(iris, p=0.5, stratify = TRUE, classes = "factor", seed = 1)

# or add NAs to numeric variables
mlim.na(iris, p=0.5, classes = "numeric", seed = 1)
}
}
\author{
E. F. Haghish
}
