% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlim.preimpute.R
\name{mlim.preimpute}
\alias{mlim.preimpute}
\title{carries out preimputation}
\usage{
mlim.preimpute(data, preimpute, seed = NULL, report = NULL, debug = FALSE)
}
\arguments{
\item{data}{data.frame with missing values}

\item{preimpute}{character. specify the algorithm for preimputation. the
supported options are "RF" (Random Forest) and "mm"
(mean-mode replacement). the default is "RF", which carries
a parallel random forest imputation, using all the CPUs available.}

\item{report}{filename. if a filename is specified, the \code{"md.log"} R
package is used to generate a Markdown progress report for the
imputation. the format of the report is adopted based on the
\code{'verbosity'} argument. the higher the verbosity, the more
technical the report becomes. if verbosity equals "debug", then
a log file is generated, which includes time stamp and shows
the function that has generated the message. otherwise, a
reduced markdown-like report is generated.
this feature is used internally by 'mlim' function.}

\item{debug.}{logical. if TRUE, the debugging feature of the report is activated.
this feature is used internally by 'mlim' function.}
}
\value{
imputed data.frame
}
\description{
instead of replacing missing data with mean and mode, a smarter
             start-point would be to use fast imputation algorithms and then
             optimize the imputed dataset with mlim. this procedure usually
             requires less iterations and will savea lot of computation
             resources.
}
\author{
E. F. Haghish
}
