% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlim.summarize.R
\name{mlim.summarize}
\alias{mlim.summarize}
\title{mlim imputation summary}
\usage{
mlim.summarize(data)
}
\arguments{
\item{data}{data.frame with missing values}

\item{preimpute}{character. specify the algorithm for preimputation. the
supported options are "RF" (Random Forest) and "mm"
(mean-mode replacement). the default is "RF", which carries
a parallel random forest imputation, using all the CPUs available.}

\item{report}{filename. if a filename is specified, the \code{"md.log"} R
package is used to generate a Markdown progress report for the
imputation. the format of the report is adopted based on the
\code{'verbosity'} argument. the higher the verbosity, the more
technical the report becomes. if verbosity equals "debug", then
a log file is generated, which includes time stamp and shows
the function that has generated the message. otherwise, a
reduced markdown-like report is generated.
this feature is used internally by 'mlim' function.}

\item{debug.}{logical. if TRUE, the debugging feature of the report is activated.
this feature is used internally by 'mlim' function.}
}
\value{
imputed data.frame
}
\description{
provides information about estimated accuracy of the imputation as well
             as the overall procedure of the imputation.
}
\author{
E. F. Haghish
}
